migrations/models/1_20230716205744_add_unique.py
from tortoise import BaseDBAsyncClient


async def upgrade(db: BaseDBAsyncClient) -> str:
    return """
        CREATE UNIQUE INDEX "uid_rates_date_9d5fa2" ON "rates" ("date", "cargo_type");"""


async def downgrade(db: BaseDBAsyncClient) -> str:
    return """
        DROP INDEX "uid_rates_date_9d5fa2";"""
